updated vision_moves to work only with radians

+++ src/init_free/init_cub_2.c
@@ -6,7 +6,7 @@
 /*   By: emonacho <emonacho@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/09/17 22:38:41 by emonacho          #+#    #+#             */
-/*   Updated: 2025/10/09 17:25:39 by emonacho         ###   ########.fr       */
+/*   Updated: 2025/10/09 17:58:00 by emonacho         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,7 +23,7 @@ int		init_parsed_data(t_main *cub)
 	//cub->map.plyr_start_pos[X] = 2;
 	//cub->map.plyr_start_pos[Y] = 2;
 	fprintf(stderr, "%s⚠️ cub->gfx.txtr_res..........: %d\n⚠️ cub->map.dim[X]............: %ld\n⚠️ cub->map.dim[Y]............: %ld%s\n", YEL, cub->gfx.txtr_res, cub->map.dim[X], cub->map.dim[Y], RESET);
-	fprintf(stderr, "%s⚠️ cub->map.plyr_start_ori....: %ld\n⚠️ cub->map.plyr_start_pos[X].: %ld\n⚠️ cub->map.plyr_start_pos[Y].: %ld%s\n", YEL, cub->map.plyr_start_ori, cub->map.plyr_start_pos[X], cub->map.plyr_start_pos[Y], RESET);
+	fprintf(stderr, "%s⚠️ cub->map.plyr_start_ori....: %lf\n⚠️ cub->map.plyr_start_pos[X].: %ld\n⚠️ cub->map.plyr_start_pos[Y].: %ld%s\n", YEL, cub->map.plyr_start_ori, cub->map.plyr_start_pos[X], cub->map.plyr_start_pos[Y], RESET);
 	//cub->map.grid = init_2d_array(cub->map.dim[Y], cub->map.dim[X]);
 	//.............................⚠️TESTS
 	//cub->plyr.pos_mp[X] = cub->map.plyr_start_pos[X];
@@ -34,7 +34,7 @@ int		init_parsed_data(t_main *cub)
 	cub->plyr.pos[Y] = (double)cub->map.plyr_start_pos[Y];
 	cub->plyr.fov_max = 16;
 	cub->plyr.fov_val = 2;
-	cub->plyr.aov = (double)cub->map.plyr_start_ori;
+	cub->plyr.aov = cub->map.plyr_start_ori;
 	return (0);
 }
